requirements.txt
APScheduler==3.6.3
atomicwrites==1.4.1
attrs==22.1.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==2.0.12
colorama==0.4.5
coloredlogs==15.0.1
flake8==3.9.2
flake8-docstrings==1.6.0
humanfriendly==10.0
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
mccabe==0.6.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycodestyle==2.7.0
pydocstyle==6.1.1
pyflakes==2.3.1
pyparsing==3.0.9
pyreadline==2.1
pytest==6.2.5
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
snowballstemmer==2.2.0
toml==0.10.2
tornado==6.2
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.11
zipp==3.8.1
